+++ src/apps/balancer-v2/helpers/balancer-v2.pool.token-helper.ts
@@ -75,9 +75,9 @@ export class BalancerV2PoolTokensHelper {
           const appToken = appTokens.find(p => p.address === tokenAddress);
           return appToken ?? baseToken;
         });
-        const tokens = _.compact(tokensRaw);
 
-        if (tokens.some(isUndefined)) return null;
+        if (tokensRaw.some(isUndefined)) return null;
+        const tokens = _.compact(tokensRaw);
 
         const reserves = tokens.map((t, i) => Number(poolTokensRaw.balances[i]) / 10 ** t.decimals);
         const liquidity = tokens.reduce((acc, v, i) => acc + v.price * reserves[i], 0);