+++ src/apps/qi-dao/index.ts
@@ -1,0 +1,1 @@
+export { QI_DAO_DEFINITION, QiDaoAppDefinition } from './qi-dao.definition';